third_party/blink/web_tests/virtual/http-cache-sql-backend/external/wpt/fetch/http-cache/credentials.tentative.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] same-origin: 2xAnonymous, 2xCredentialled, 1xAnonymous
  assert_equals: Response 3 header Server-Request-Count is "1", not "2" expected "2" but got "1"
[FAIL] same-origin: 2xCredentialled, 2xAnonymous, 1xCredentialled
  assert_equals: Response 3 header Server-Request-Count is "1", not "2" expected "2" but got "1"
Harness: the test ran to completion.
